requirements.txt
fastapi
uvicorn
newspaper4k
feedparser
transformers
torch
rake-nltk
nltk
gtts
requests
beautifulsoup4
streamlit
transformers
torch
newspaper3k
requests
beautifulsoup4
rake-nltk
nltk
gtts
flask
gradio
